Extracted some field check handling functions out of ApplicationController into a standalone object.
