+++ src/test/scala/views/html/EventObjectivesFormSpec.scala
@@ -31,7 +31,6 @@ class EventObjectivesFormSpec extends WordSpecLike with Matchers with OptionValu
 
 
     sectionForm(
-      Application(ApplicationId(1), ApplicationFormId(1)),
       ApplicationOverview(ApplicationId(1), ApplicationFormId(1), Seq()),
       ApplicationForm(ApplicationFormId(1), OpportunityId(1), Seq(ApplicationFormSection(5, "Event Audience"))),
       section,
