+++ src/test/scala/views/html/TopicsAndSpeakersFormSpec.scala
@@ -30,7 +30,6 @@ class TopicsAndSpeakersFormSpec extends WordSpecLike with Matchers with OptionVa
     val name = "topicAndSpeakerRules"
 
     sectionForm(
-      Application(ApplicationId(1), ApplicationFormId(1)),
       ApplicationOverview(ApplicationId(1), ApplicationFormId(1), Seq()),
       ApplicationForm(ApplicationFormId(1), OpportunityId(1), Seq(ApplicationFormSection(5, "Event Audience"))),
       section,
